chore: move adhoc pages to its own dir
