+++ webapp/javascript/index.jsx
@@ -16,9 +16,9 @@ import ContinuousDiffView from './pages/ContinuousDiffView';
 import Continuous from './components/Continuous';
 import Settings from './components/Settings';
 import Sidebar from './components/Sidebar';
-import AdhocSingle from './pages/AdhocSingle';
-import AdhocComparison from './pages/AdhocComparison';
-import AdhocDiff from './pages/AdhocDiff';
+import AdhocSingle from './pages/adhoc/AdhocSingle';
+import AdhocComparison from './pages/adhoc/AdhocComparison';
+import AdhocDiff from './pages/adhoc/AdhocDiff';
 import ServiceDiscoveryApp from './pages/ServiceDiscovery';
 import ServerNotifications from './components/ServerNotifications';
 import Protected from './components/Protected';
